Experimental: * extract context for each link and index it * context retrieval uses:   - StringMap for efficient hashmap serialization   - ObjectCache a simple FIFO cache for caching context fields * experiment with various scoring schemes, use related as boost for   sloppy phrase matches, rank as boost for exact phrases - probably   will be changed futher
Devel:
* drop link_analysis index, maintain an index of pagelinks and redirect
  tables in links index
* fix bug 11103, use FSUtils for all filesystem related operations
* localization: read meta namespaces names per dbname, pass iid to
  wikitokenizer, etc.. 
* RelatedBuilder can build related mapping from both links index and
  from dump, cleanup and field compression
* new search method: related (returns the related mapping for article)


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/lucene-search-2.1@26333 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ src/org/wikimedia/lsearch/related/CompactRelated.java
@@ -36,7 +36,7 @@ public class CompactRelated {
 		this.title = title;
 	}
 	public String serialize(){
-		return score+" "+relates;
+		return ((float)score)+" "+relates;
 	}
 	
 	public static ArrayList<String> convertToStringList(Collection<CompactRelated> rel){
